Move filter options fetch to dashboard
